// std-toolkit/src/db/dynamodb/client/generated/types.ts
interface _AttributeValue {
  /**
   * An attribute of type String. For example:
   * "S": "Hello"
   */
  S?: string;
  /**
   * An attribute of type Number. For example:
   * "N": "123.45"
   * Numbers are sent across the network to DynamoDB as strings, to maximize compatibility
   * across languages and libraries. However, DynamoDB treats them as number type attributes
   * for mathematical operations.
   */
  N?: string;
  /**
   * An attribute of type Binary. For example:
   * "B": "dGhpcyB0ZXh0IGlzIGJhc2U2NC1lbmNvZGVk"
   */
  B?: Uint8Array | string;
  /**
   * An attribute of type String Set. For example:
   * "SS": ["Giraffe", "Hippo" ,"Zebra"]
   */
  SS?: Array<string>;
  /**
   * An attribute of type Number Set. For example:
   * "NS": ["42.2", "-19", "7.5", "3.14"]
   * Numbers are sent across the network to DynamoDB as strings, to maximize compatibility
   * across languages and libraries. However, DynamoDB treats them as number type attributes
   * for mathematical operations.
   */
  NS?: Array<string>;
  /**
   * An attribute of type Binary Set. For example:
   * "BS": ["U3Vubnk=", "UmFpbnk=", "U25vd3k="]
   */
  BS?: Array<Uint8Array | string>;
  /**
   * An attribute of type Map. For example:
   * "M": {"Name": {"S": "Joe"}, "Age": {"N": "35"}}
   */
  M?: Record<string, AttributeValue>;
  /**
   * An attribute of type List. For example:
   * "L": [ {"S": "Cookies"} , {"S": "Coffee"}, {"N": "3.14159"}]
   */
  L?: Array<AttributeValue>;
  /**
   * An attribute of type Null. For example:
   * "NULL": true
   */
  NULL?: boolean;
  /**
   * An attribute of type Boolean. For example:
   * "BOOL": true
   */
  BOOL?: boolean;
}

export type AttributeValue =
  | (_AttributeValue & { S: string })
  | (_AttributeValue & { N: string })
  | (_AttributeValue & { B: Uint8Array | string })
  | (_AttributeValue & { SS: Array<string> })
  | (_AttributeValue & { NS: Array<string> })
  | (_AttributeValue & { BS: Array<Uint8Array | string> })
  | (_AttributeValue & { M: Record<string, AttributeValue> })
  | (_AttributeValue & { L: Array<AttributeValue> })
  | (_AttributeValue & { NULL: boolean })
  | (_AttributeValue & { BOOL: boolean });
// ...
export interface Capacity {
  /**
   * The total number of read capacity units consumed on a table or an index.
   */
  ReadCapacityUnits?: number;
  /**
   * The total number of write capacity units consumed on a table or an index.
   */
  WriteCapacityUnits?: number;
  /**
   * The total number of capacity units consumed on a table or an index.
   */
  CapacityUnits?: number;
}
// ...
export interface ConsumedCapacity {
  /**
   * The name of the table that was affected by the operation. If you had specified the
   * Amazon Resource Name (ARN) of a table in the input, you'll see the table ARN in the response.
   */
  TableName?: string;
  /**
   * The total number of capacity units consumed by the operation.
   */
  CapacityUnits?: number;
  /**
   * The total number of read capacity units consumed by the operation.
   */
  ReadCapacityUnits?: number;
  /**
   * The total number of write capacity units consumed by the operation.
   */
  WriteCapacityUnits?: number;
  /**
   * The amount of throughput consumed on the table affected by the operation.
   */
  Table?: Capacity;
  /**
   * The amount of throughput consumed on each local index affected by the
   * operation.
   */
  LocalSecondaryIndexes?: Record<string, Capacity>;
  /**
   * The amount of throughput consumed on each global index affected by the
   * operation.
   */
  GlobalSecondaryIndexes?: Record<string, Capacity>;
}
// ...
export type GlobalSecondaryIndexes = Array<GlobalSecondaryIndexInfo>;
// ...
export type LocalSecondaryIndexes = Array<LocalSecondaryIndexInfo>;
// ...
export interface ScanOutput {
  /**
   * An array of item attributes that match the scan criteria. Each element in this array
   * consists of an attribute name and the value for that attribute.
   */
  Items?: Array<Record<string, AttributeValue>>;
  /**
   * The number of items in the response.
   * If you set ScanFilter in the request, then Count is the
   * number of items returned after the filter was applied, and ScannedCount is
   * the number of matching items before the filter was applied.
   * If you did not use a filter in the request, then Count is the same as
   * ScannedCount.
   */
  Count?: number;
  /**
   * The number of items evaluated, before any ScanFilter is applied. A high
   * ScannedCount value with few, or no, Count results
   * indicates an inefficient Scan operation. For more information, see Count and
   * ScannedCount in the Amazon DynamoDB Developer
   * Guide.
   * If you did not use a filter in the request, then ScannedCount is the same
   * as Count.
   */
  ScannedCount?: number;
  /**
   * The primary key of the item where the operation stopped, inclusive of the previous
   * result set. Use this value to start a new operation, excluding this value in the new
   * request.
   * If LastEvaluatedKey is empty, then the "last page" of results has been
   * processed and there is no more data to be retrieved.
   * If LastEvaluatedKey is not empty, it does not necessarily mean that there
   * is more data in the result set. The only way to know when you have reached the end of
   * the result set is when LastEvaluatedKey is empty.
   */
  LastEvaluatedKey?: Record<string, AttributeValue>;
  /**
   * The capacity units consumed by the Scan operation. The data returned
   * includes the total provisioned throughput consumed, along with statistics for the table
   * and any indexes involved in the operation. ConsumedCapacity is only
   * returned if the ReturnConsumedCapacity parameter was specified. For more
   * information, see Capacity unit consumption for read operations in the Amazon
   * DynamoDB Developer Guide.
   */
  ConsumedCapacity?: ConsumedCapacity;
}
// ...
export type TableName = string;